w3c-reports/12092014/mac/runned-tests/websockets/interfaces/WebSocket/events/020-expected.txt

CONSOLE MESSAGE: WebSocket network error: Proxy connection could not be established, unexpected response code

Harness Error. harness_status.status = 2 , harness_status.message = null
